perturbation_toolbox/requirements.txt
pillow
audiomentations
numpy
torch
torchvision
scikit-image
Wand
opencv-python
soundfile
transformers
accelerate
pyroomacoustics